Refactored CarouselLayout.js to be a shared component.
